refactor: move file drop overlay support
