+++ Sources/ContentView.swift
@@ -7,590 +7,6 @@ import ObjectiveC
 import UniformTypeIdentifiers
 import WebKit
 
-// MARK: - File Drop Overlay
-
-enum DragOverlayRoutingPolicy {
-    static let bonsplitTabTransferType = NSPasteboard.PasteboardType("com.splittabbar.tabtransfer")
-    static let sidebarTabReorderType = NSPasteboard.PasteboardType(SidebarTabDragPayload.typeIdentifier)
-
-    static func hasBonsplitTabTransfer(_ pasteboardTypes: [NSPasteboard.PasteboardType]?) -> Bool {
-        guard let pasteboardTypes else { return false }
-        return pasteboardTypes.contains(bonsplitTabTransferType)
-    }
-
-    static func hasSidebarTabReorder(_ pasteboardTypes: [NSPasteboard.PasteboardType]?) -> Bool {
-        guard let pasteboardTypes else { return false }
-        return pasteboardTypes.contains(sidebarTabReorderType)
-    }
-
-    static func hasFileURL(_ pasteboardTypes: [NSPasteboard.PasteboardType]?) -> Bool {
-        guard let pasteboardTypes else { return false }
-        return pasteboardTypes.contains(.fileURL)
-    }
-
-    static func shouldCaptureFileDropDestination(
-        pasteboardTypes: [NSPasteboard.PasteboardType]?,
-        hasLocalDraggingSource: Bool
-    ) -> Bool {
-        // Local file drags (e.g. in-app draggable folder views) are valid drop
-        // inputs; rely on explicit non-file drag types below to avoid hijacking
-        // Bonsplit/sidebar drags.
-        _ = hasLocalDraggingSource
-        guard hasFileURL(pasteboardTypes) else { return false }
-
-        // Prefer explicit non-file drag types so stale fileURL entries cannot hijack
-        // Bonsplit tab drags or sidebar tab reorder drags.
-        if hasBonsplitTabTransfer(pasteboardTypes) { return false }
-        if hasSidebarTabReorder(pasteboardTypes) { return false }
-        return true
-    }
-
-    static func shouldCaptureFileDropDestination(
-        pasteboardTypes: [NSPasteboard.PasteboardType]?
-    ) -> Bool {
-        shouldCaptureFileDropDestination(
-            pasteboardTypes: pasteboardTypes,
-            hasLocalDraggingSource: false
-        )
-    }
-
-    static func shouldCaptureFileDropOverlay(
-        pasteboardTypes: [NSPasteboard.PasteboardType]?,
-        eventType: NSEvent.EventType?
-    ) -> Bool {
-        guard shouldCaptureFileDropDestination(pasteboardTypes: pasteboardTypes) else { return false }
-        guard isDragMouseEvent(eventType) else { return false }
-        return true
-    }
-
-    static func shouldCaptureSidebarExternalOverlay(
-        hasSidebarDragState: Bool,
-        pasteboardTypes: [NSPasteboard.PasteboardType]?
-    ) -> Bool {
-        guard hasSidebarDragState else { return false }
-        return hasSidebarTabReorder(pasteboardTypes)
-    }
-
-    static func shouldCaptureSidebarExternalOverlay(
-        draggedTabId: UUID?,
-        pasteboardTypes: [NSPasteboard.PasteboardType]?
-    ) -> Bool {
-        shouldCaptureSidebarExternalOverlay(
-            hasSidebarDragState: draggedTabId != nil,
-            pasteboardTypes: pasteboardTypes
-        )
-    }
-
-    static func shouldPassThroughPortalHitTesting(
-        pasteboardTypes: [NSPasteboard.PasteboardType]?,
-        eventType: NSEvent.EventType?
-    ) -> Bool {
-        guard isPortalDragEvent(eventType) else { return false }
-        return hasBonsplitTabTransfer(pasteboardTypes) || hasSidebarTabReorder(pasteboardTypes)
-    }
-
-    private static func isDragMouseEvent(_ eventType: NSEvent.EventType?) -> Bool {
-        eventType == .leftMouseDragged
-            || eventType == .rightMouseDragged
-            || eventType == .otherMouseDragged
-    }
-
-    private static func isPortalDragEvent(_ eventType: NSEvent.EventType?) -> Bool {
-        // Restrict portal pass-through to explicit drag-motion events so stale
-        // NSPasteboard(name: .drag) types cannot hijack normal pointer input.
-        guard let eventType else { return false }
-        switch eventType {
-        case .leftMouseDragged, .rightMouseDragged, .otherMouseDragged:
-            return true
-        default:
-            return false
-        }
-    }
-}
-
-/// Transparent NSView installed on the window's theme frame (above the NSHostingView) to
-/// handle file/URL drags from Finder. Nested NSHostingController layers (created by bonsplit's
-/// SinglePaneWrapper) prevent AppKit's NSDraggingDestination routing from reaching deeply
-/// embedded terminal views. This overlay sits above the entire content view hierarchy and
-/// intercepts file drags, forwarding drops to the GhosttyNSView under the cursor.
-///
-/// Mouse events are forwarded to the views below via a hide-send-unhide pattern so clicks,
-/// scrolls, and other interactions pass through normally.
-final class FileDropOverlayView: NSView {
-    /// Fallback handler when no terminal is found under the drop point.
-    var onDrop: (([URL]) -> Bool)?
-    private var isForwardingMouseEvent = false
-    private weak var forwardedMouseDragTarget: NSView?
-    private var forwardedMouseDragButton: ForwardedMouseDragButton?
-    /// The WKWebView currently receiving forwarded drag events, so we can
-    /// synthesize draggingExited/draggingEntered as the cursor moves.
-    private weak var activeDragWebView: WKWebView?
-    /// The WKWebView that accepted prepareForDragOperation so conclude can be
-    /// delivered to the same browser target after the drop completes.
-    private weak var preparedDragWebView: WKWebView?
-    private var lastHitTestLogSignature: String?
-    private var lastDragRouteLogSignatureByPhase: [String: String] = [:]
-
-    override var acceptsFirstResponder: Bool { false }
-
-    override init(frame frameRect: NSRect) {
-        super.init(frame: frameRect)
-        registerForDraggedTypes([.fileURL])
-    }
-
-    required init?(coder: NSCoder) { fatalError("init(coder:) not implemented") }
-
-    private enum ForwardedMouseDragButton: Equatable {
-        case left
-        case right
-        case other(Int)
-    }
-
-    private func dragButton(for event: NSEvent) -> ForwardedMouseDragButton? {
-        switch event.type {
-        case .leftMouseDown, .leftMouseUp, .leftMouseDragged:
-            return .left
-        case .rightMouseDown, .rightMouseUp, .rightMouseDragged:
-            return .right
-        case .otherMouseDown, .otherMouseUp, .otherMouseDragged:
-            return .other(Int(event.buttonNumber))
-        default:
-            return nil
-        }
-    }
-
-    private func shouldTrackForwardedMouseDragStart(for eventType: NSEvent.EventType) -> Bool {
-        switch eventType {
-        case .leftMouseDown, .rightMouseDown, .otherMouseDown:
-            return true
-        default:
-            return false
-        }
-    }
-
-    private func shouldTrackForwardedMouseDragEnd(for eventType: NSEvent.EventType) -> Bool {
-        switch eventType {
-        case .leftMouseUp, .rightMouseUp, .otherMouseUp:
-            return true
-        default:
-            return false
-        }
-    }
-
-    // MARK: Hit-testing — participation is routed by DragOverlayRoutingPolicy so
-    // file-drop, bonsplit tab drags, and sidebar tab reorder drags cannot conflict.
-
-    override func hitTest(_ point: NSPoint) -> NSView? {
-        let pb = NSPasteboard(name: .drag)
-        let eventType = NSApp.currentEvent?.type
-        let shouldCapture = DragOverlayRoutingPolicy.shouldCaptureFileDropOverlay(
-            pasteboardTypes: pb.types,
-            eventType: eventType
-        )
-#if DEBUG
-        logHitTestDecision(
-            pasteboardTypes: pb.types,
-            eventType: eventType,
-            shouldCapture: shouldCapture
-        )
-#endif
-        guard shouldCapture else { return nil }
-
-        return super.hitTest(point)
-    }
-
-    // MARK: Mouse forwarding — safety net for the rare case where stale drag pasteboard
-    // data causes hitTest to return self when no drag is actually active.
-    // We hit-test contentView directly and dispatch to the target rather than using
-    // window.sendEvent(), which caches the mouse target and causes infinite recursion.
-
-    private func forwardEvent(_ event: NSEvent) {
-        guard !isForwardingMouseEvent else { return }
-        guard let window, let contentView = window.contentView else { return }
-        let eventButton = dragButton(for: event)
-
-        isForwardingMouseEvent = true
-        isHidden = true
-        defer {
-            isHidden = false
-            isForwardingMouseEvent = false
-        }
-
-        let target: NSView?
-        if let eventButton,
-           forwardedMouseDragButton == eventButton,
-           let activeTarget = forwardedMouseDragTarget,
-           activeTarget.window != nil {
-            // Preserve normal AppKit mouse-delivery semantics: once a drag starts,
-            // keep routing dragged/up events to the original mouseDown target.
-            target = activeTarget
-        } else {
-            let point = contentView.convert(event.locationInWindow, from: nil)
-            target = contentView.hitTest(point)
-        }
-
-        guard let target, target !== self else {
-            if shouldTrackForwardedMouseDragEnd(for: event.type),
-               let eventButton,
-               forwardedMouseDragButton == eventButton {
-                forwardedMouseDragTarget = nil
-                forwardedMouseDragButton = nil
-            }
-            return
-        }
-
-        if shouldTrackForwardedMouseDragStart(for: event.type), let eventButton {
-            forwardedMouseDragTarget = target
-            forwardedMouseDragButton = eventButton
-        }
-
-        switch event.type {
-        case .leftMouseDown: target.mouseDown(with: event)
-        case .leftMouseUp: target.mouseUp(with: event)
-        case .leftMouseDragged: target.mouseDragged(with: event)
-        case .rightMouseDown: target.rightMouseDown(with: event)
-        case .rightMouseUp: target.rightMouseUp(with: event)
-        case .rightMouseDragged: target.rightMouseDragged(with: event)
-        case .otherMouseDown: target.otherMouseDown(with: event)
-        case .otherMouseUp: target.otherMouseUp(with: event)
-        case .otherMouseDragged: target.otherMouseDragged(with: event)
-        case .scrollWheel: target.scrollWheel(with: event)
-        default: break
-        }
-
-        if shouldTrackForwardedMouseDragEnd(for: event.type),
-           let eventButton,
-           forwardedMouseDragButton == eventButton {
-            forwardedMouseDragTarget = nil
-            forwardedMouseDragButton = nil
-        }
-    }
-
-    override func mouseDown(with event: NSEvent) { forwardEvent(event) }
-    override func mouseUp(with event: NSEvent) { forwardEvent(event) }
-    override func mouseDragged(with event: NSEvent) { forwardEvent(event) }
-    override func rightMouseDown(with event: NSEvent) { forwardEvent(event) }
-    override func rightMouseUp(with event: NSEvent) { forwardEvent(event) }
-    override func rightMouseDragged(with event: NSEvent) { forwardEvent(event) }
-    override func otherMouseDown(with event: NSEvent) { forwardEvent(event) }
-    override func otherMouseUp(with event: NSEvent) { forwardEvent(event) }
-    override func otherMouseDragged(with event: NSEvent) { forwardEvent(event) }
-    override func scrollWheel(with event: NSEvent) { forwardEvent(event) }
-
-    // MARK: NSDraggingDestination – accept file drops over terminal and browser views.
-    //
-    // AppKit sends draggingEntered once when the drag enters this overlay, then
-    // draggingUpdated as the cursor moves within it. We track which WKWebView (if
-    // any) is under the cursor and synthesize enter/exit calls so the browser's
-    // HTML5 drag events (dragenter, dragleave, drop) fire correctly.
-
-    override func draggingEntered(_ sender: any NSDraggingInfo) -> NSDragOperation {
-        return updateDragTarget(sender, phase: "entered")
-    }
-
-    override func draggingUpdated(_ sender: any NSDraggingInfo) -> NSDragOperation {
-        return updateDragTarget(sender, phase: "updated")
-    }
-
-    override func draggingExited(_ sender: (any NSDraggingInfo)?) {
-        preparedDragWebView = nil
-        if let prev = activeDragWebView {
-            prev.draggingExited(sender)
-            activeDragWebView = nil
-        }
-    }
-
-    override func prepareForDragOperation(_ sender: any NSDraggingInfo) -> Bool {
-        let hasLocalDraggingSource = sender.draggingSource != nil
-        let types = sender.draggingPasteboard.types
-        let shouldCapture = DragOverlayRoutingPolicy.shouldCaptureFileDropDestination(
-            pasteboardTypes: types,
-            hasLocalDraggingSource: hasLocalDraggingSource
-        )
-        let webView = shouldCapture ? (activeDragWebView ?? webViewUnderPoint(sender.draggingLocation)) : nil
-        let terminal = terminalUnderPoint(sender.draggingLocation)
-        let hasTerminalTarget = terminal != nil
-#if DEBUG
-        logDragRouteDecision(
-            phase: "prepare",
-            pasteboardTypes: types,
-            shouldCapture: shouldCapture,
-            hasLocalDraggingSource: hasLocalDraggingSource,
-            hasTerminalTarget: hasTerminalTarget
-        )
-#endif
-        guard shouldCapture else {
-            preparedDragWebView = nil
-            return false
-        }
-        if let webView {
-            preparedDragWebView = webView
-            return webView.prepareForDragOperation(sender)
-        }
-        preparedDragWebView = nil
-        return hasTerminalTarget
-    }
-
-    override func performDragOperation(_ sender: any NSDraggingInfo) -> Bool {
-        let hasLocalDraggingSource = sender.draggingSource != nil
-        let types = sender.draggingPasteboard.types
-        let shouldCapture = DragOverlayRoutingPolicy.shouldCaptureFileDropDestination(
-            pasteboardTypes: types,
-            hasLocalDraggingSource: hasLocalDraggingSource
-        )
-        let webView = shouldCapture
-            ? (preparedDragWebView ?? activeDragWebView ?? webViewUnderPoint(sender.draggingLocation))
-            : nil
-        let terminal = terminalUnderPoint(sender.draggingLocation)
-        let hasTerminalTarget = terminal != nil
-#if DEBUG
-        logDragRouteDecision(
-            phase: "perform",
-            pasteboardTypes: types,
-            shouldCapture: shouldCapture,
-            hasLocalDraggingSource: hasLocalDraggingSource,
-            hasTerminalTarget: hasTerminalTarget
-        )
-#endif
-        guard shouldCapture else {
-            preparedDragWebView = nil
-            activeDragWebView = nil
-            return false
-        }
-        if let webView {
-            preparedDragWebView = webView
-            return webView.performDragOperation(sender)
-        }
-        preparedDragWebView = nil
-        activeDragWebView = nil
-        guard let terminal else { return false }
-        return terminal.performDragOperation(sender)
-    }
-
-    override func concludeDragOperation(_ sender: (any NSDraggingInfo)?) {
-        defer {
-            preparedDragWebView = nil
-            activeDragWebView = nil
-        }
-        guard let sender else { return }
-        let webView = preparedDragWebView ?? activeDragWebView ?? webViewUnderPoint(sender.draggingLocation)
-        webView?.concludeDragOperation(sender)
-    }
-
-    private func updateDragTarget(_ sender: any NSDraggingInfo, phase: String) -> NSDragOperation {
-        let loc = sender.draggingLocation
-        let hasLocalDraggingSource = sender.draggingSource != nil
-        let types = sender.draggingPasteboard.types
-        let shouldCapture = DragOverlayRoutingPolicy.shouldCaptureFileDropDestination(
-            pasteboardTypes: types,
-            hasLocalDraggingSource: hasLocalDraggingSource
-        )
-        let webView = shouldCapture ? webViewUnderPoint(loc) : nil
-
-        if let prev = activeDragWebView, prev !== webView {
-            prev.draggingExited(sender)
-            activeDragWebView = nil
-        }
-
-        if let webView {
-            if activeDragWebView !== webView {
-                activeDragWebView = webView
-                return webView.draggingEntered(sender)
-            }
-            return webView.draggingUpdated(sender)
-        }
-
-        let hasTerminalTarget = terminalUnderPoint(loc) != nil
-#if DEBUG
-        logDragRouteDecision(
-            phase: phase,
-            pasteboardTypes: types,
-            shouldCapture: shouldCapture,
-            hasLocalDraggingSource: hasLocalDraggingSource,
-            hasTerminalTarget: hasTerminalTarget
-        )
-#endif
-        guard shouldCapture, hasTerminalTarget else { return [] }
-        return .copy
-    }
-
-    private func debugPasteboardTypes(_ types: [NSPasteboard.PasteboardType]?) -> String {
-        guard let types, !types.isEmpty else { return "-" }
-        return types.map(\.rawValue).joined(separator: ",")
-    }
-
-    /// Hit-tests the window to find a WKWebView (browser panel) under the cursor.
-    func webViewUnderPoint(_ windowPoint: NSPoint) -> WKWebView? {
-        if let window,
-           let portalWebView = BrowserWindowPortalRegistry.webViewAtWindowPoint(windowPoint, in: window) {
-            return portalWebView
-        }
-
-        guard let window, let contentView = window.contentView else { return nil }
-        isHidden = true
-        defer { isHidden = false }
-        let point = contentView.convert(windowPoint, from: nil)
-        let hitView = contentView.hitTest(point)
-
-        var current: NSView? = hitView
-        while let view = current {
-            if let webView = view as? WKWebView { return webView }
-            current = view.superview
-        }
-        return nil
-    }
-
-    private func debugTopHitViewForCurrentEvent() -> String {
-        guard let window,
-              let currentEvent = NSApp.currentEvent,
-              let contentView = window.contentView,
-              let themeFrame = contentView.superview else { return "-" }
-
-        let pointInTheme = themeFrame.convert(currentEvent.locationInWindow, from: nil)
-        // Don't toggle isHidden here — it triggers setNeedsDisplay which can
-        // exceed AppKit's display-pass limit during cursor-update display cycles.
-        guard let hit = themeFrame.hitTest(pointInTheme) else { return "nil" }
-        var chain: [String] = []
-        var current: NSView? = hit
-        var depth = 0
-        while let view = current, depth < 6 {
-            chain.append(debugHitViewDescriptor(view))
-            current = view.superview
-            depth += 1
-        }
-        return chain.joined(separator: "->")
-    }
-
-    private func debugHitViewDescriptor(_ view: NSView) -> String {
-        let className = String(describing: type(of: view))
-        let ptr = String(describing: Unmanaged.passUnretained(view).toOpaque())
-        let dragTypes = debugRegisteredDragTypes(view)
-        return "\(className)@\(ptr){dragTypes=\(dragTypes)}"
-    }
-
-    private func debugRegisteredDragTypes(_ view: NSView) -> String {
-        let types = view.registeredDraggedTypes
-        guard !types.isEmpty else { return "-" }
-
-        let interestingTypes = types.filter { type in
-            let raw = type.rawValue
-            return raw == NSPasteboard.PasteboardType.fileURL.rawValue
-                || raw == DragOverlayRoutingPolicy.bonsplitTabTransferType.rawValue
-                || raw == DragOverlayRoutingPolicy.sidebarTabReorderType.rawValue
-                || raw.contains("public.text")
-                || raw.contains("public.url")
-                || raw.contains("public.data")
-        }
-        let selected = interestingTypes.isEmpty ? Array(types.prefix(3)) : interestingTypes
-        let rendered = selected.map(\.rawValue).joined(separator: ",")
-        if selected.count < types.count {
-            return "\(rendered),+\(types.count - selected.count)"
-        }
-        return rendered
-    }
-
-    private func hasRelevantDragTypes(_ types: [NSPasteboard.PasteboardType]?) -> Bool {
-        guard let types else { return false }
-        return types.contains(.fileURL)
-            || types.contains(DragOverlayRoutingPolicy.bonsplitTabTransferType)
-            || types.contains(DragOverlayRoutingPolicy.sidebarTabReorderType)
-    }
-
-    private func debugEventName(_ eventType: NSEvent.EventType?) -> String {
-        guard let eventType else { return "none" }
-        switch eventType {
-        case .cursorUpdate: return "cursorUpdate"
-        case .appKitDefined: return "appKitDefined"
-        case .systemDefined: return "systemDefined"
-        case .applicationDefined: return "applicationDefined"
-        case .periodic: return "periodic"
-        case .mouseMoved: return "mouseMoved"
-        case .mouseEntered: return "mouseEntered"
-        case .mouseExited: return "mouseExited"
-        case .flagsChanged: return "flagsChanged"
-        case .leftMouseDown: return "leftMouseDown"
-        case .leftMouseUp: return "leftMouseUp"
-        case .leftMouseDragged: return "leftMouseDragged"
-        case .rightMouseDown: return "rightMouseDown"
-        case .rightMouseUp: return "rightMouseUp"
-        case .rightMouseDragged: return "rightMouseDragged"
-        case .otherMouseDown: return "otherMouseDown"
-        case .otherMouseUp: return "otherMouseUp"
-        case .otherMouseDragged: return "otherMouseDragged"
-        case .scrollWheel: return "scrollWheel"
-        default: return "other(\(eventType.rawValue))"
-        }
-    }
-
-#if DEBUG
-    private func logHitTestDecision(
-        pasteboardTypes: [NSPasteboard.PasteboardType]?,
-        eventType: NSEvent.EventType?,
-        shouldCapture: Bool
-    ) {
-        let isDragEvent = eventType == .leftMouseDragged
-            || eventType == .rightMouseDragged
-            || eventType == .otherMouseDragged
-        guard shouldCapture || isDragEvent || hasRelevantDragTypes(pasteboardTypes) else { return }
-
-        let signature = "\(shouldCapture ? 1 : 0)|\(debugEventName(eventType))|\(debugPasteboardTypes(pasteboardTypes))"
-        guard lastHitTestLogSignature != signature else { return }
-        lastHitTestLogSignature = signature
-        dlog(
-            "overlay.fileDrop.hitTest capture=\(shouldCapture ? 1 : 0) " +
-            "event=\(debugEventName(eventType)) " +
-            "topHit=\(debugTopHitViewForCurrentEvent()) " +
-            "types=\(debugPasteboardTypes(pasteboardTypes))"
-        )
-    }
-
-    private func logDragRouteDecision(
-        phase: String,
-        pasteboardTypes: [NSPasteboard.PasteboardType]?,
-        shouldCapture: Bool,
-        hasLocalDraggingSource: Bool,
-        hasTerminalTarget: Bool
-    ) {
-        guard shouldCapture || hasRelevantDragTypes(pasteboardTypes) else { return }
-        let signature = [
-            shouldCapture ? "1" : "0",
-            hasLocalDraggingSource ? "1" : "0",
-            hasTerminalTarget ? "1" : "0",
-            debugPasteboardTypes(pasteboardTypes)
-        ].joined(separator: "|")
-        guard lastDragRouteLogSignatureByPhase[phase] != signature else { return }
-        lastDragRouteLogSignatureByPhase[phase] = signature
-        dlog(
-            "overlay.fileDrop.\(phase) capture=\(shouldCapture ? 1 : 0) " +
-            "localSource=\(hasLocalDraggingSource ? 1 : 0) " +
-            "hasTerminal=\(hasTerminalTarget ? 1 : 0) " +
-            "types=\(debugPasteboardTypes(pasteboardTypes))"
-        )
-    }
-#endif
-    /// Hit-tests the window to find the GhosttyNSView under the cursor.
-    func terminalUnderPoint(_ windowPoint: NSPoint) -> GhosttyNSView? {
-        if let window,
-           let portalTerminal = TerminalWindowPortalRegistry.terminalViewAtWindowPoint(windowPoint, in: window) {
-            return portalTerminal
-        }
-
-        guard let window, let contentView = window.contentView else { return nil }
-        isHidden = true
-        defer { isHidden = false }
-        let point = contentView.convert(windowPoint, from: nil)
-        let hitView = contentView.hitTest(point)
-
-        var current: NSView? = hitView
-        while let view = current {
-            if let terminal = view as? GhosttyNSView { return terminal }
-            current = view.superview
-        }
-        return nil
-    }
-}
-
-var fileDropOverlayKey: UInt8 = 0
 private var commandPaletteWindowOverlayKey: UInt8 = 0
 private var tmuxWorkspacePaneWindowOverlayKey: UInt8 = 0
 let commandPaletteOverlayContainerIdentifier = NSUserInterfaceItemIdentifier("cmux.commandPalette.overlay.container")
@@ -1398,87 +814,6 @@ enum MountedWorkspacePresentationPolicy {
             isRenderedVisible: isRenderedVisible,
             isPanelVisible: isRenderedVisible,
             renderOpacity: renderOpacity
-        )
-    }
-}
-
-/// Installs a FileDropOverlayView on the window's theme frame for Finder file drag support.
-private func findFileDropOverlayView(in root: NSView?) -> FileDropOverlayView? {
-    guard let root else { return nil }
-    if let overlay = root as? FileDropOverlayView {
-        return overlay
-    }
-    for subview in root.subviews {
-        if let overlay = findFileDropOverlayView(in: subview) {
-            return overlay
-        }
-    }
-    return nil
-}
-
-private func configureFileDropOverlay(_ overlay: FileDropOverlayView, tabManager: TabManager) {
-    overlay.onDrop = { [weak tabManager] urls in
-        MainActor.assumeIsolated {
-            guard let tabManager, let terminal = tabManager.selectedWorkspace?.focusedTerminalPanel else { return false }
-            return terminal.hostedView.handleDroppedURLs(urls)
-        }
-    }
-}
-
-private func attachFileDropOverlay(_ overlay: FileDropOverlayView, to contentView: NSView, in themeFrame: NSView) {
-    overlay.translatesAutoresizingMaskIntoConstraints = false
-    themeFrame.addSubview(overlay, positioned: .above, relativeTo: contentView)
-    NSLayoutConstraint.activate([
-        overlay.topAnchor.constraint(equalTo: contentView.topAnchor),
-        overlay.bottomAnchor.constraint(equalTo: contentView.bottomAnchor),
-        overlay.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
-        overlay.trailingAnchor.constraint(equalTo: contentView.trailingAnchor)
-    ])
-}
-
-@discardableResult
-func installFileDropOverlay(on window: NSWindow, tabManager: TabManager) -> Bool {
-    guard let contentView = window.contentView,
-          let themeFrame = contentView.superview else { return false }
-
-    let existingOverlay =
-        (objc_getAssociatedObject(window, &fileDropOverlayKey) as? FileDropOverlayView)
-        ?? findFileDropOverlayView(in: themeFrame)
-
-    if let existingOverlay {
-        configureFileDropOverlay(existingOverlay, tabManager: tabManager)
-        objc_setAssociatedObject(window, &fileDropOverlayKey, existingOverlay, .OBJC_ASSOCIATION_RETAIN)
-        guard existingOverlay.superview !== themeFrame else { return true }
-        existingOverlay.removeFromSuperview()
-        attachFileDropOverlay(existingOverlay, to: contentView, in: themeFrame)
-        return true
-    }
-
-    let overlay = FileDropOverlayView(frame: contentView.frame)
-    configureFileDropOverlay(overlay, tabManager: tabManager)
-    // Publish the overlay before mutating the view tree so any re-entrant lookup resolves
-    // the in-flight view instead of installing a second overlay during layout.
-    objc_setAssociatedObject(window, &fileDropOverlayKey, overlay, .OBJC_ASSOCIATION_RETAIN)
-    attachFileDropOverlay(overlay, to: contentView, in: themeFrame)
-    return true
-}
-
-private func installFileDropOverlayWhenReady(
-    on window: NSWindow,
-    tabManager: TabManager,
-    remainingAttempts: Int = 16
-) {
-    guard !installFileDropOverlay(on: window, tabManager: tabManager),
-          remainingAttempts > 0 else { return }
-
-    // Defer retrying until the next main-loop turn so we don't mutate the
-    // NSThemeFrame hierarchy while SwiftUI/AppKit is still attaching views.
-    DispatchQueue.main.async { [weak window, weak tabManager] in
-        guard let window, let tabManager else { return }
-        installFileDropOverlayWhenReady(
-            on: window,
-            tabManager: tabManager,
-            remainingAttempts: remainingAttempts - 1
         )
     }
 }
@@ -13742,7 +13077,7 @@ private final class SidebarDragAutoScrollController: ObservableObject {
     }
 }
 
-private enum SidebarTabDragPayload {
+enum SidebarTabDragPayload {
     static let typeIdentifier = "com.cmux.sidebar-tab-reorder"
     static let dropContentType = UTType(exportedAs: typeIdentifier)
     static let dropContentTypes: [UTType] = [dropContentType]
